more replacement of some hard coded int32_t

+++ modules/core/bessel/include/nt2/toolbox/bessel/functions/scalar/jni.hpp
@@ -82,7 +82,7 @@ namespace nt2 { namespace ext
     NT2_FUNCTOR_CALL(2)
     {
       result_type x = a1;
-      const int32_t n1 = nt2::abs(a0);
+      const size_t n1 = nt2::abs(a0);
       result_type sign = a0<Zero<A0>()?cospi(n1):One<A0>();
       if( n1 == 0 )
         return( sign * nt2::j0(x) );
@@ -92,7 +92,7 @@ namespace nt2 { namespace ext
         return mul(sign, (mul(Two<result_type>(), nt2::j1(x) / x)  -  nt2::j0(x)) );
 
       /* continued fraction */
-      int k = 24;
+      ptrdiff_t k = 24;
       result_type pk = Two<result_type>()*(result_type(n1) + result_type(k));
       result_type ans = pk;
       result_type xk = sqr(x);
